# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bookshelves API",
        "description": "Documentation about Bookshelves API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/authors/{author}": {
            "get": {
                "tags": [
                    "authors"
                ],
                "summary": "Show author by author slug",
                "description": "Get details for a single author with list of series and books, check /authors endpoint to get list of slugs",
                "operationId": "findAuthorByAuthorSlug",
                "parameters": [
                    {
                        "name": "author",
                        "in": "path",
                        "description": "Slug of author name like 'auel-jean-m' for Jean M. Auel",
                        "required": true,
                        "style": "form",
                        "example": "auel-jean-m"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Invalid author-slug value or book-slug value",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
        "/download/book/{author}/{book}": {
            "get": {
                "tags": [
                    "download"
                ],
                "summary": "Download specific book",
                "description": "Download specific book, check /books endpoint to get list of slugs",
                "operationId": "App\\Http\\Controllers\\Api\\DownloadController::book",
                "parameters": [
                    {
                        "name": "author",
                        "in": "path",
                        "description": "Slug of author name like 'auel-jean-m' for Jean M. Auel",
                        "required": true,
                        "style": "form",
                        "example": "auel-jean-m"
                    },
                    {
                        "name": "book",
                        "in": "path",
                        "description": "Slug of book name like 'les-refuges-de-pierre' for Les refuges de pierre",
                        "required": true,
                        "style": "form",
                        "example": "les-refuges-de-pierre"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/download/serie/{author}/{serie}": {
            "get": {
                "tags": [
                    "download"
                ],
                "summary": "Download specific serie's books with ZIP",
                "description": "Download specific serie, check /series endpoint to get list of slugs",
                "operationId": "App\\Http\\Controllers\\Api\\DownloadController::serie",
                "parameters": [
                    {
                        "name": "author",
                        "in": "path",
                        "description": "Slug of author name like 'auel-jean-m' for Jean M. Auel",
                        "required": true,
                        "style": "form",
                        "example": "auel-jean-m"
                    },
                    {
                        "name": "book",
                        "in": "path",
                        "description": "Slug of book name like 'les-enfants-de-la-terre' for Les enfants de la terre",
                        "required": true,
                        "style": "form",
                        "example": "les-enfants-de-la-terre"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/download/author/{author}": {
            "get": {
                "tags": [
                    "download"
                ],
                "summary": "Download specific author's books with ZIP",
                "description": "Download specific author, check /authors endpoint to get list of slugs",
                "operationId": "App\\Http\\Controllers\\Api\\DownloadController::author",
                "parameters": [
                    {
                        "name": "author",
                        "in": "path",
                        "description": "Slug of author name like 'auel-jean-m' for Jean M. Auel",
                        "required": true,
                        "style": "form",
                        "example": "auel-jean-m"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
        "/series": {
            "get": {
                "tags": [
                    "series"
                ],
                "summary": "List of series",
                "description": "Series",
                "operationId": "App\\Http\\Controllers\\Api\\SerieController::index",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Integer to choose how many books you show in each page",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 32
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/series/{author}/{series}": {
            "get": {
                "tags": [
                    "series"
                ],
                "summary": "Show series by author slug and by series slug",
                "description": "Get details for a single series with list of books, check /series endpoint to get list of slugs",
                "operationId": "findAuthorByAuthorSlug",
                "parameters": [
                    {
                        "name": "author",
                        "in": "path",
                        "description": "Slug of author name like 'lovecraft-howard-phillips' for Howard Phillips Lovecraft",
                        "required": true,
                        "style": "form",
                        "example": "lovecraft-howard-phillips"
                    },
                    {
                        "name": "series",
                        "in": "path",
                        "description": "Slug of book name like 'cthulhu-le-mythe-fr' for Cthulhu : Le Mythe",
                        "required": true,
                        "style": "form",
                        "example": "cthulhu-le-mythe-fr"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Invalid author-slug value or book-slug value",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        }
    },
...
    "security": [
        []
    ]
}